requirements.txt
aiohttp==3.8.5 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
aiosignal==1.3.1 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
annotated-types==0.5.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
async-timeout==4.0.3 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
attrs==23.1.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
blinker==1.6.2 ; python_full_version >= "3.9.0" and python_version < "4"
certifi==2023.7.22 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
charset-normalizer==3.2.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
click==8.1.7 ; python_full_version >= "3.9.0" and python_version < "4"
cloudevents==1.9.0 ; python_full_version >= "3.9.0" and python_version < "4"
colorama==0.4.6 ; python_full_version >= "3.9.0" and python_version < "4" and platform_system == "Windows"
deprecation==2.1.0 ; python_full_version >= "3.9.0" and python_version < "4"
flask==2.3.3 ; python_full_version >= "3.9.0" and python_version < "4"
frozenlist==1.4.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
functions-framework==3.4.0 ; python_full_version >= "3.9.0" and python_version < "4"
gunicorn==20.1.0 ; python_full_version >= "3.9.0" and python_version < "4" and platform_system != "Windows"
idna==3.4 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
importlib-metadata==6.8.0 ; python_full_version >= "3.9.0" and python_version < "3.10"
itsdangerous==2.1.2 ; python_full_version >= "3.9.0" and python_version < "4"
jinja2==3.1.2 ; python_full_version >= "3.9.0" and python_version < "4"
markupsafe==2.1.3 ; python_full_version >= "3.9.0" and python_version < "4"
multidict==6.0.4 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
openai==0.27.10 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
packaging==23.1 ; python_full_version >= "3.9.0" and python_version < "4"
pydantic-core==2.6.3 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
pydantic==2.3.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
python-dotenv==1.0.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
requests==2.31.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
setuptools==68.1.2 ; python_full_version >= "3.9.0" and python_version < "4" and platform_system != "Windows"
tabulate==0.9.0 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
termcolor==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.66.1 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
typing-extensions==4.7.1 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
urllib3==2.0.4 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
watchdog==3.0.0 ; python_full_version >= "3.9.0" and python_version < "4"
werkzeug==2.3.7 ; python_full_version >= "3.9.0" and python_version < "4"
yarl==1.9.2 ; python_full_version >= "3.9.0" and python_full_version < "4.0.0"
yaspin==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.16.2 ; python_full_version >= "3.9.0" and python_version < "3.10"
textbase-client>=0.1.2